requirements.txt
requests
tqdm
pydantic>=2.4
feedparser
beautifulsoup4
lxml
pyhumps